feat: handle special errors differently for rate limiting and automatic log out
