+++ src/home/fetch-github/fetch-issues-preview.ts
@@ -1,9 +1,10 @@
 import { RequestError } from "@octokit/request-error";
 import { Octokit } from "@octokit/rest";
 import { getGitHubAccessToken, getGitHubUserName } from "../getters/get-github-access-token";
-import { getGitHubUser } from "../getters/get-github-user";
 import { GitHubIssue } from "../github-types";
-import { displayPopupMessage, renderErrorInModal } from "../rendering/display-popup-modal";
+import { displayPopupMessage } from "../rendering/display-popup-modal";
+import { gitHubLoginButton } from "../rendering/render-github-login-button";
+import { handleRateLimit } from "./handle-rate-limit";
 import { TaskNoFull } from "./preview-to-full-mapping";
 
 async function checkPrivateRepoAccess(): Promise<boolean> {
@@ -58,20 +59,7 @@ export async function fetchIssuePreviews(): Promise<TaskNoFull[]> {
 
     // Fetch issues from the private repository only if the user has access
     if (hasPrivateRepoAccess) {
-      const { data: privateResponse } = await octokit.issues.listForRepo({
-        owner: "ubiquity",
-        repo: "devpool-directory-private",
-        state: "open",
-      });
-      const privateIssues = privateResponse.filter((issue: GitHubIssue) => !issue.pull_request);
-
-      // Mark private issues
-      const privateIssuesWithFlag = privateIssues.map((issue) => {
-        return issue;
-      });
-
-      // Combine public and private issues
-      freshIssues = [...privateIssuesWithFlag, ...publicIssues];
+      await fetchPrivateIssues(publicIssues);
     } else {
       // If user doesn't have access, only load issues from the public repository
       freshIssues = publicIssues;
@@ -80,7 +68,11 @@ export async function fetchIssuePreviews(): Promise<TaskNoFull[]> {
     if (error instanceof RequestError && error.status === 403) {
       await handleRateLimit(octokit, error);
     } else {
-      renderErrorInModal(error as Error, "Error fetching issue previews");
+      // renderErrorInModal(error as Error, "You have been rate limited. Please login to increase your limits."); // @DEV: user another method to render the modal not as an error
+      displayPopupMessage({ modalHeader: "GitHub API rate limit exceeded.", modalBody: "You have been rate limited. Please login to increase your limits.", isError: false });
+      gitHubLoginButton?.classList.add("highlight");
+      // throw error;
+      // console.error("You have been rate limited. Please login to increase your limits. ", error);
     }
   }
 
@@ -92,45 +84,25 @@ export async function fetchIssuePreviews(): Promise<TaskNoFull[]> {
   })) as TaskNoFull[];
 
   return tasks;
+
+  async function fetchPrivateIssues(publicIssues: GitHubIssue[]) {
+    const { data: privateResponse } = await octokit.issues.listForRepo({
+      owner: "ubiquity",
+      repo: "devpool-directory-private",
+      state: "open",
+    });
+    const privateIssues = privateResponse.filter((issue: GitHubIssue) => !issue.pull_request);
+
+    // Mark private issues
+    const privateIssuesWithFlag = privateIssues.map((issue) => {
+      return issue;
+    });
+
+    // Combine public and private issues
+    freshIssues = [...privateIssuesWithFlag, ...publicIssues];
+  }
 }
 
-function rateLimitModal(message: string) {
+export function rateLimitModal(message: string) {
   displayPopupMessage({ modalHeader: `GitHub API rate limit exceeded.`, modalBody: message, isError: false });
-}
-
-type RateLimit = {
-  reset: number | null;
-  user: boolean;
-};
-
-export async function handleRateLimit(octokit?: Octokit, error?: RequestError) {
-  const rate: RateLimit = {
-    reset: null,
-    user: false,
-  };
-
-  if (error?.response?.headers["x-ratelimit-reset"]) {
-    rate.reset = parseInt(error.response.headers["x-ratelimit-reset"]);
-  }
-
-  if (octokit) {
-    try {
-      const core = await octokit.rest.rateLimit.get();
-      const remaining = core.data.resources.core.remaining;
-      const reset = core.data.resources.core.reset;
-
-      rate.reset = !rate.reset && remaining === 0 ? reset : rate.reset;
-      rate.user = (await getGitHubUser()) ? true : false;
-    } catch (err) {
-      renderErrorInModal(err as Error, "Error handling GitHub rate limit");
-    }
-  }
-
-  const resetParsed = rate.reset && new Date(rate.reset * 1000).toLocaleTimeString();
-
-  if (!rate.user) {
-    rateLimitModal(`You have been rate limited. Please log in to GitHub to increase your GitHub API limits, otherwise you can try again at ${resetParsed}.`);
-  } else {
-    rateLimitModal(`You have been rate limited. Please try again at ${resetParsed}.`);
-  }
 }
